project/requirements.txt
asyncpg
fastapi
pytest
python-dotenv
python-multipart
requests
tortoise-orm
uvicorn